Move feature tests to test folder
