+++ docs/design/TODO.md
@@ -1,0 +1,8 @@
+cleanup code
+compile feature-based agent state
+petting zoo corporation
+step function for grid (S0 initially)
+    - step for individual -> update internal state (some are part of the global state S1)
+    - update_state for global state (S1 -> S2)
+    - populate global state to individuals (S2 to local)
+
