+++ powergrid/envs/single_agent/cigre_mv.py
@@ -5,7 +5,7 @@ import pandapower as pp
 from os.path import dirname, abspath
 from collections import OrderedDict
 
-from powergrid.base_env import GridBaseEnv
+from deprecated.base_env import GridBaseEnv
 from powergrid.devices import *
 
 def read_data(train, load_area, renew_area, price_area):
