+++ powergrid/envs/single_agent/ieee13_mg.py
@@ -5,7 +5,7 @@ import pandapower as pp
 from os.path import dirname, abspath
 from collections import OrderedDict
 
-from powergrid.base_env import GridBaseEnv
+from deprecated.base_env import GridBaseEnv
 from powergrid.networks.ieee13 import IEEE13Bus
 from powergrid.devices import *
