+++ tests/core/features/test_status_block.py
