+++ tests/core/features/test_thermal_loading.py
